services/cnn-inference/requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.40.0
numpy==2.4.2
onnxruntime==1.24.2
pillow==12.1.1
pydantic==2.12.5
